feat(tooling-explorer): Rename file to NetworkSelector
