+++ apps/explorer/src/components/network/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './WrappedNetworkSelect';
+export * from './NetworkSelector';
